Close #90: Simplification of 'nav' data
Don't use Statocles's built-in 'nav' site-constructor argument, since
that bails the build complaining about missing 'href' attributes.

At the same time remove the 'title' element which isn't used for anything
in our site and doesn't provide for a meaningful 'title' attribute on
the A tag.

Then, rename the 'name' key to 'title' as that seems more appropriate
as a name for a link.